+++ backend-doc/Memory.md
@@ -2,6 +2,6 @@ Pseudo backend that holds all data in memory.
 
 ```
 backend {
-  type = akashic.storage.backend.Memory
+  type = akashic.storage.backend.impl.Memory
 }
 ```
