+++ src/main/resources/reference.conf
@@ -1,6 +1,6 @@
 akashic.storage {
   backend {
-    type = akashic.storage.backend.Local
+    type = akashic.storage.backend.impl.Local
     mountpoint = /tmp/akashic-storage-test
   }
