+++ src/test/resources/backend-local.conf
@@ -1,4 +1,4 @@
 backend {
-  type = akashic.storage.backend.Local
+  type = akashic.storage.backend.impl.Local
   mountpoint = /tmp/l
 }
