+++ src/test/resources/backend-memory.conf
@@ -1,3 +1,3 @@
 backend {
-  type = akashic.storage.backend.Memory
+  type = akashic.storage.backend.impl.Memory
 }
